+++ src/plugins/wizard/public/application/components/data_tab/config_panel.scss
