+++ src/plugins/wizard/public/application/components/data_tab/config_panel.tsx
@@ -5,8 +5,8 @@
 
 import { EuiForm } from '@elastic/eui';
 import React from 'react';
-import { useVisualizationType } from '../../../utils/use';
-import { useTypedSelector } from '../../../utils/state_management';
+import { useVisualizationType } from '../../utils/use';
+import { useTypedSelector } from '../../utils/state_management';
 import './config_panel.scss';
 import { mapSchemaToAggPanel } from './utils/schema_to_dropbox';
 import { SecondaryPanel } from './secondary_panel';
