+++ src/plugins/wizard/public/application/components/data_tab/field_selector.scss
@@ -1,11 +1,13 @@
-@import "../../../util";
+@import "../../util";
 
 .wizFieldSelector {
   @include scrollNavParent(auto 1fr);
+
   padding: $euiSizeS;
 
   &__fieldGroups {
     @include euiYScrollWithShadows;
+
     overflow-y: auto;
     margin-right: -$euiSizeS;
     padding-right: $euiSizeS;
@@ -14,7 +16,7 @@
 
   &__fieldGroup {
     margin-top: $euiSizeS;
-  
+
     &:first-child {
       margin-top: 0;
     }
