+++ src/plugins/wizard/public/application/components/data_tab/field_selector.tsx
@@ -11,12 +11,12 @@ import {
   IndexPatternField,
   OPENSEARCH_FIELD_TYPES,
   OSD_FIELD_TYPES,
-} from '../../../../../../data/public';
+} from '../../../../../data/public';
 import { FieldSelectorField } from './field_selector_field';
 
 import './field_selector.scss';
-import { useTypedSelector } from '../../../utils/state_management';
-import { useIndexPattern } from '../../../utils/use';
+import { useTypedSelector } from '../../utils/state_management';
+import { useIndexPattern } from '../../utils/use';
 
 interface IFieldCategories {
   categorical: IndexPatternField[];
