+++ src/plugins/wizard/public/application/components/data_tab/field_selector_field.scss
@@ -1,5 +1,6 @@
 .wizFieldSelectorField {
   @include euiBottomShadowSmall;
+
   padding: $euiSizeXS;
   background-color: $euiColorEmptyShade;
   border: $euiBorderThin;
