+++ src/plugins/wizard/public/application/components/data_tab/index.scss
@@ -1,7 +1,8 @@
-@import "../../../util";
+@import "../../util";
 
 .wizDataTab {
   @include scrollNavParent;
+
   display: grid;
   grid-template-columns: 50% 50%;
 }
