+++ src/plugins/wizard/public/application/components/data_tab/items/dropbox.scss
