+++ src/plugins/wizard/public/application/components/data_tab/items/index.tsx
@@ -3,7 +3,5 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export * from './types';
 export * from './dropbox';
 export { Title } from './title';
-export { FormField } from './form_field';
