+++ src/plugins/wizard/public/application/components/data_tab/items/use/index.ts
@@ -2,5 +2,5 @@
  * Copyright OpenSearch Contributors
  * SPDX-License-Identifier: Apache-2.0
  */
-export * from './containers';
-export * from './constants';
+
+export { useDropbox } from './use_dropbox';
