+++ src/plugins/wizard/public/application/components/data_tab/items/use/use_dropbox.tsx
@@ -5,21 +5,21 @@
 
 import { useState, useEffect, useCallback, useMemo } from 'react';
 import { cloneDeep } from 'lodash';
-import { BucketAggType, IndexPatternField, propFilter } from '../../../../../../../../data/common';
-import { Schema } from '../../../../../../../../vis_default_editor/public';
-import { FieldDragDataType } from '../../../../../utils/drag_drop/types';
-import { useTypedDispatch, useTypedSelector } from '../../../../../utils/state_management';
+import { BucketAggType, IndexPatternField, propFilter } from '../../../../../../../data/common';
+import { Schema } from '../../../../../../../vis_default_editor/public';
+import { FieldDragDataType } from '../../../../utils/drag_drop/types';
+import { useTypedDispatch, useTypedSelector } from '../../../../utils/state_management';
 import { DropboxState, DropboxDisplay } from '../types';
 import { DropboxProps } from '../dropbox';
-import { useDrop } from '../../../../../utils/drag_drop';
+import { useDrop } from '../../../../utils/drag_drop';
 import {
   editAgg,
   reorderAgg,
   updateAggConfigParams,
-} from '../../../../../utils/state_management/visualization_slice';
-import { useIndexPattern } from '../../../../../../application/utils/use/use_index_pattern';
-import { useOpenSearchDashboards } from '../../../../../../../../opensearch_dashboards_react/public';
-import { WizardServices } from '../../../../../../types';
+} from '../../../../utils/state_management/visualization_slice';
+import { useIndexPattern } from '../../../../utils/use/use_index_pattern';
+import { useOpenSearchDashboards } from '../../../../../../../opensearch_dashboards_react/public';
+import { WizardServices } from '../../../../../types';
 
 const filterByName = propFilter('name');
 const filterByType = propFilter('type');
