--- src/plugins/wizard/public/application/contributions/constants.ts
@@ -1,12 +1,0 @@
-/*
- * Copyright OpenSearch Contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-import { CommonItemTypes } from './containers/common/items';
-import { DataTabItemTypes } from './containers/data_tab/items';
-
-export const ItemTypes = {
-  ...CommonItemTypes,
-  ...DataTabItemTypes,
-};
