--- src/plugins/wizard/public/application/contributions/containers/common/items/index.tsx
@@ -1,9 +1,0 @@
-/*
- * Copyright OpenSearch Contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-export * from './types';
-
-export { Select } from './select';
-export { TextInput } from './text_input';
