--- src/plugins/wizard/public/application/contributions/containers/data_tab/items/use/index.ts
@@ -1,7 +1,0 @@
-/*
- * Copyright OpenSearch Contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-export { useDropbox } from './use_dropbox';
-export { useFormField } from './use_form_field';
