--- src/plugins/wizard/public/application/contributions/containers/index.ts
@@ -1,7 +1,0 @@
-/*
- * Copyright OpenSearch Contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-export * from './data_tab';
-export * from './style_tab';
