--- src/plugins/wizard/public/application/contributions/containers/style_tab/index.tsx
@@ -1,12 +1,0 @@
-/*
- * Copyright OpenSearch Contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-import React from 'react';
-
-export const STYLE_TAB_ID = 'style_tab';
-
-export const StyleTab = () => {
-  return <div>TODO: Layout styles come here.</div>;
-};
